asoc: kona-smith: Fix FLL timeouts
FLL timeouts are seen in certain usecases (Voice calls)
because we try to lock the FLL before the bus clock is turned on.
Make sure the TDM clock for Sen tdm interface is always on before
trying to lock the FLL on the codec.

Change-Id: I4882aa73d886d65d6368a217adcffbb14076028b
Signed-off-by: Ravi Vembu <raviv@motorola.com>
Mot-CRs-Fixed: (CR)
Reviewed-on: https://gerrit.mot.com/1862081
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Vladimir Karpovich <vkarp@motorola.com>
Reviewed-by: Nelson Guedes Paulo Junior <npaulo@motorola.com>
Reviewed-by: Luiz Eduardo Mendes Matheus <luizmmat@motorola.com>
Reviewed-by: Ling Jin <lingjin@motorola.com>
Submit-Approved: Jira Key