dwc3: gadget: Handle Missed Isoc packets
Increase the tx fifo size for Isoc endpoints when the burst size
specified is > 6. Also log the failed isoc packets and do not stop
the active transfer unless we run out of requests. This should
give a chance for the failed packets to be requeued on the bus.

Signed-off-by: Vijayavardhan Vennapusa <vvreddy@qti.qualcomm.com>
Signed-off-by: Ravi Vembu <raviv@motorola.com>

Mot-CRs-fixed: (CR)
Change-Id: I1db98b64f6a9395e75ffe525999356e23ab421ab
Reviewed-on: https://gerrit.mot.com/1968338
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Ling Jin <lingjin@motorola.com>
Reviewed-by: Daniel Vacura <w36195@motorola.com>
Reviewed-by: Jian Luan <jamesluan@motorola.com>
Submit-Approved: Jira Key